Cast the median to numeric so get_practice_outcomes can return a row
The function created cleanly and then failed on every call, so the migration
looked applied and the panel stayed empty:

  42804 structure of query does not match function result type
  DETAIL: Returned type double precision does not match expected type numeric
          in column 15.

percentile_cont has no numeric variant -- it is WITHIN GROUP (ORDER BY double
precision) -> double precision -- so ordering by the numeric progress_pct casts
up to double, and single-argument round(double) stays double. Column 15 is
declared numeric, and RETURN QUERY checks the tuple structure strictly.

Swept the other three migrations for the same class of mismatch. Every other
round() is the two-argument round(numeric, int) form, which exists only for
numeric, so all of them return what they declare. This was the only one.

Worth noting for next time: a migration running without error proves the
function PARSES, not that it RUNS. Return-type mismatches surface at call time
only, which is exactly how this survived a clean run.

+++ supabase/add_practice_outcomes.sql
@@ -147,7 +147,18 @@ begin
       -- percentile_cont matches the JS median in both parities: it returns the
       -- middle value for odd counts and the mean of the two middles for even,
       -- which is exactly what snapshot.mjs computes.
-      (select round(percentile_cont(0.5) within group (order by a.progress_pct))
+      --
+      -- The ::numeric is REQUIRED, not tidiness. percentile_cont has no numeric
+      -- variant -- it is WITHIN GROUP (ORDER BY double precision) -> double
+      -- precision -- so ordering by a numeric column casts up to double, and
+      -- round(double) returns double. This column is declared numeric, and
+      -- RETURN QUERY checks the tuple structure strictly:
+      --   42804 structure of query does not match function result type
+      --   DETAIL: Returned type double precision does not match expected type
+      --           numeric in column 15.
+      -- The function CREATEs fine and fails only when called, so a clean
+      -- migration run proves nothing here.
+      (select round(percentile_cont(0.5) within group (order by a.progress_pct)::numeric)
          from ab_rounds a where a.progress_pct is not null)             as median_progress_pct
     from (select 1) _
   )
